requirements.txt
torch>=1.10.0
transformers>=4.18.0
datasets>=2.0.0
rouge_score>=0.1.2
nltk>=3.6.5
matplotlib>=3.5.1
pandas>=1.3.5
pyyaml>=6.0
tqdm>=4.62.3
py7zr>=0.22.0